# src/main/resources/openapi_napolya.yaml
openapi: 3.0.3
info:
  title: API na_polya 3.0
  version: 1.0.11
  description: Все необходимые API для сайта На_поля
...
paths:
...
  /queues:
    get:
      tags:
        - Queue
      summary: Получить список id очередей
      responses:
        '200':
          description: Успешный ответ с массивом id очередей
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
                  description: Список id очередей
                  example: [1, 2, 3]
    post:
      tags:
        - Queue
      summary: Создать новую очередь
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Queue'
      responses:
        '201':
          description: Очередь успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Queue'
  /queues/{id}:
    get:
      tags:
        - Queue
      summary: Получить информацию об очереди
      parameters:
        - name: id
          in: path
          required: true
          description: ID очереди
          schema:
            type: integer
      responses:
        '200':
          description: Информация об очереди
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Queue'
        '404':
          description: Очередь не найдена

    put:
      tags:
        - Queue
      summary: Обновить информацию об очереди
      parameters:
        - name: id
          in: path
          required: true
          description: ID очереди
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Queue'
      responses:
        '200':
          description: Очередь успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Queue'
        '404':
          description: Очередь не найдена
    
    delete:
      tags:
        - Queue
      summary: Удалить очередь
      parameters:
        - name: id
          in: path
          required: true
          description: ID очереди
          schema:
            type: integer
      responses:
        '204':
          description: Очередь успешно удалена
        '404':
          description: Очередь не найдена
  /providers:
    get:
      tags:
        - Provider
      summary: Получить список названий провайдеров
      responses:
        '200':
          description: Успешный ответ с массивом названий провайдеров
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  description: Список названий провайдеров
                  example: ["Поляна", "Интерспорт"]
    post:
      tags:
        - Provider
      summary: Создать нового провайдера
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Provider'
      responses:
        '201':
          description: Провайдер успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
  /providers/{id}:
    get:
      tags:
        - Provider
      summary: Получить информацию о провайдере
      parameters:
        - name: id
          in: path
          required: true
          description: ID провайдера
          schema:
            type: integer
      responses:
        '200':
          description: Информация о провайдере
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '404':
          description: Провайдер не найден

    put:
      tags:
        - Provider
      summary: Обновить информацию о провайдере
      parameters:
        - name: id
          in: path
          required: true
          description: ID провайдера
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Provider'
      responses:
        '200':
          description: Провайдер успешно обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '404':
          description: Провайдер не найден
    
    delete:
      tags:
        - Provider
      summary: Удалить провайдера
      parameters:
        - name: id
          in: path
          required: true
          description: ID провайдера
          schema:
            type: integer
      responses:
        '204':
          description: Провайдер успешно удалён
        '404':
          description: Провайдер не найден
...
components:
  schemas:
...
    Queue: # Определение сущности "Очередь"
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор очереди
          example: 1
        players_ids:
          type: array
          description: id игроков, находящихся в очереди
          items:
            type: integer
          example: [5, 6, 7]
        game_id:
          type: integer
          description: id игры, на которую стоит очередь
          example: 1

      required:
        - id
        - players_ids
        - game_id
    Provider: # Определение сущности "Провайдер"
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор провайдера
          example: 1
        name:
          type: string
          description: Название провайдера
          example: "Поляна"
        fields_ids:
          type: array
          description: id полей, которые предоставляет провайдер
          items:
            type: integer
          example: [1, 2, 3]
      
      required:
        - id
        - name
...